chore: 기존 VC와 View 분리되던 방식을 VC로 통합 및 내부 구현 방식 변경
